Confidence region for model parameters

Computes boundary points of a two-dimensional confidence region for model parameters.

Usage

# S4 method for EL
confreg(
  object,
  parm,
  level = 0.95,
  cv = NULL,
  npoints = 50L,
  control = el_control()
)

Arguments

object

An object that inherit from class EL, including CEL, LM, and GLM.

parm

A specification of which parameters are to be given a confidence region, either a vector of numbers or a vector of names. It should be a vector of length two of the form c(x, y). If missing, the first two parameter in object are considered.

level

A single numeric for the confidence level required. Defaults to 0.95. It is ignored if cv is non-NULL.

cv

A single numeric for the critical value for calibration of empirical likelihood ratio statistic. Defaults to NULL and set to qchisq(level, 2L). It must be compatible with the th value in control.

npoints

A single integer for the number of boundary points to compute. Defaults to 50.

control

An object of class ControlEL constructed by el_control().

Value

An object of class ConfregEL.

References

Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.” The Annals of Statistics 18 (1): 90–120. doi:10.1214/aos/1176347494 .

See also

Examples

fit <- el_lm(mpg ~ ., data = mtcars)
confreg(fit, parm = c(2, 3), level = 0.95, cv = qchisq(0.95, 2))
#> An object of class "ConfregEL"
#> Slot "points":
#>              [,1]        [,2]
#>  [1,] -0.42678529 0.013335240
#>  [2,] -0.16784831 0.006209274
#>  [3,] -0.14138448 0.005646927
#>  [4,] -0.13127516 0.005482134
#>  [5,] -0.12588240 0.005395725
#>  [6,] -0.12244339 0.005341155
#>  [7,] -0.11999490 0.005302105
#>  [8,] -0.11811030 0.005272809
#>  [9,] -0.11657249 0.005248477
#> [10,] -0.11525536 0.005228197
#> [11,] -0.11408071 0.005209443
#> [12,] -0.11299374 0.005192730
#> [13,] -0.11195372 0.005177427
#> [14,] -0.11092621 0.005161246
#> [15,] -0.10987810 0.005144954
#> [16,] -0.10877372 0.005127813
#> [17,] -0.10756996 0.005109979
#> [18,] -0.10620728 0.005089045
#> [19,] -0.10459801 0.005064121
#> [20,] -0.10260103 0.005034443
#> [21,] -0.09993529 0.004976233
#> [22,] -0.09607175 0.004886210
#> [23,] -0.08972953 0.004739268
#> [24,] -0.07679663 0.004440211
#> [25,] -0.03701314 0.003932881
#> [26,]  0.41320342 0.013335240
#> [27,] -0.04541215 0.021676557
#> [28,] -0.07680309 0.022228609
#> [29,] -0.08840125 0.022457116
#> [30,] -0.09463265 0.022575419
#> [31,] -0.09862985 0.022642707
#> [32,] -0.10146855 0.022699505
#> [33,] -0.10366489 0.022734306
#> [34,] -0.10544742 0.022778788
#> [35,] -0.10698362 0.022806544
#> [36,] -0.10835731 0.022824255
#> [37,] -0.10962452 0.022854820
#> [38,] -0.11084064 0.022869345
#> [39,] -0.11204217 0.022898851
#> [40,] -0.11326769 0.022913837
#> [41,] -0.11456048 0.022937630
#> [42,] -0.11596839 0.022957553
#> [43,] -0.11756895 0.022992170
#> [44,] -0.11945314 0.023020885
#> [45,] -0.12180102 0.023064440
#> [46,] -0.12491540 0.023125345
#> [47,] -0.12935645 0.023184631
#> [48,] -0.13670888 0.023339708
#> [49,] -0.15142986 0.023602767
#> [50,] -0.19954542 0.024465483
#> 
#> Slot "estimates":
#>         cyl        disp 
#> -0.11144048  0.01333524 
#> 
#> Slot "level":
#> [1] NA
#> 
#> Slot "cv":
#> [1] 5.991465
#> 
#> Slot "pnames":
#> [1] "cyl"  "disp"
#>