Pairwise comparisons for general block designs

Tests all pairwise comparisons or comparisons with control for general block designs. Two single step asymptotic \(k\)-FWER (generalized family-wise error rate) controlling procedures are available: asymptotic Monte Carlo (AMC) and nonparametric bootstrap (NB).

This function is deprecated and will be removed in a future release.

Usage

el_pairwise(
  formula,
  data,
  control = NULL,
  k = 1L,
  alpha = 0.05,
  method = c("AMC", "NB"),
  B,
  nthreads = 1L,
  maxit = 10000L,
  abstol = 1e-08
)

Arguments

formula

An object of class formula (or one that can be coerced to that class) for a symbolic description of the model to be fitted. It must specify variables for response, treatment, and block as 'response ~ treatment | block'. Note that the use of vertical bar (|) separating treatment and block.

data

A data frame, list or environment (or object coercible by as.data.frame() to a data frame) containing the variables in formula.

control

An optional single character that specifies the treatment for comparisons with control.

k

A single integer for \(k\) in \(k\)-FWER. Defaults to 1.

alpha

A single numeric for the overall significance level. Defaults to 0.05.

method

A single character for the procedure to be used; either AMC or NB is supported. Defaults to AMC.

B

A single integer for the number of Monte Carlo samples for the AMC (number of bootstrap replicates for the NB).

nthreads

A single integer for the number of threads for parallel computation via OpenMP (if available). Defaults to 1.

maxit

A single integer for the maximum number of iterations for constrained minimization of empirical likelihood. Defaults to 10000.

abstol

A single numeric for the the absolute convergence tolerance for optimization. Defaults to 1e-08.

Value

A list with class c("pairwise", "melt").

References

Kim E, MacEachern S, Peruggia M (2021). “Empirical Likelihood for the Analysis of Experimental Designs.” arxiv:2112.09206. URL https://arxiv.org/abs/2112.09206.

Examples

if (FALSE) {
# All pairwise comparisons
data("clothianidin")
el_pairwise(clo ~ trt | blk, data = clothianidin, B = 1000)

# Comparisons with control
el_pairwise(clo ~ trt | blk, data = clothianidin, control = "Naked",
  method = "NB", B = 500
)}