Empirical likelihood for the standard deviation

Computes empirical likelihood for the standard deviation.

Usage

el_sd(x, mean, sd, weights = NULL, control = el_control())

Arguments

x

A numeric vector, or an object that can be coerced to a numeric vector.

mean

A single numeric for the (known) mean value.

sd

A positive single numeric for the parameter value to be tested.

weights

An optional numeric vector of weights to be used in the fitting process. The length of the vector must be the same as the length of x. Defaults to NULL, corresponding to identical weights. If non-NULL, weighted empirical likelihood is computed.

control

An object of class ControlEL constructed by el_control().

Value

An object of class EL.

Details

Let \(X_i\) be independent and identically random variable from an unknown distribution \(P\) for \(i = 1, \dots, n\). We assume that \({\textrm{E}[X_i]} = {\mu_0}\) is known and that \(P\) has a variance \(\sigma_0^2\). Given a value of \(\sigma\), the (profile) empirical likelihood ratio is defined by $$R(\sigma) = \max_{p_i}\left\{\prod_{i = 1}^n np_i : \sum_{i = 1}^n p_i (X_i - \mu_0)^2 = \sigma^2, p_i \geq 0, \sum_{i = 1}^n p_i = 1 \right\}.$$ el_sd() computes the empirical log-likelihood ratio statistic \(-2\log R(\sigma)\), along with other values in EL.

See also

Examples

set.seed(4097)
x <- rnorm(100, mean = -2, sd = 3)
w <- rep(c(1, 2), each = 50)
el_sd(x, mean = -2, sd = 3.5, weights = w)
#> 
#> 	Weighted Empirical Likelihood
#> 
#> Model: sd 
#> 
#> Maximum EL estimates:
#> [1] 3.247
#> 
#> Chisq: 1.08, df: 1, Pr(>Chisq): 0.2986
#> 
#> EL evaluation: converged 
#>