Significance tests

Extracts the results of significance tests from a model.

# S4 method for SummaryLM
sigTests(object, ...)

# S4 method for LM
sigTests(object, ...)

Arguments

object

An object that inherits from LM or SummaryLM.

...

Further arguments passed to methods.

Value

The form of the value returned by sigTests() depends on the class of its argument.

Methods (by class)

sigTests(SummaryLM): Extracts a matrix with the results of significance tests.

sigTests(LM): Extracts a list with the optimization results of significance tests.

See also

Examples

data("mtcars")
fit <- el_lm(mpg ~ ., data = mtcars)
sigTests(fit)
#> $statistic
#> (Intercept)         cyl        disp          hp        drat          wt 
#> 442.0394887   0.2334334 409.8798759 420.7490737 405.1301030 439.0974196 
#>        qsec          vs          am        gear        carb 
#> 442.4803698   0.4249416  26.6281283 401.7291427   0.5095781 
#> 
#> $iterations
#> (Intercept)         cyl        disp          hp        drat          wt 
#>           0          36           0           0           0           0 
#>        qsec          vs          am        gear        carb 
#>           0          36         200           0         115 
#> 
#> $convergence
#> (Intercept)         cyl        disp          hp        drat          wt 
#>       FALSE        TRUE       FALSE       FALSE       FALSE       FALSE 
#>        qsec          vs          am        gear        carb 
#>       FALSE        TRUE       FALSE       FALSE        TRUE 
#> 
sigTests(summary(fit))
#> Warning: `sigTests()` is deprecated for signature 'SummaryLM' in melt v1.9.1.
#>                Estimate       Chisq   Pr(>Chisq)
#> (Intercept) 12.30337416 442.0394887 3.895625e-98
#> cyl         -0.11144048   0.2334334 6.289896e-01
#> disp         0.01333524 409.8798759 3.892910e-91
#> hp          -0.02148212 420.7490737 1.676493e-93
#> drat         0.78711097 405.1301030 4.209139e-90
#> wt          -3.71530393 439.0974196 1.701700e-97
#> qsec         0.82104075 442.4803698 3.123385e-98
#> vs           0.31776281   0.4249416 5.144815e-01
#> am           2.52022689  26.6281283 2.466269e-07
#> gear         0.65541302 401.7291427 2.314856e-89
#> carb        -0.19941925   0.5095781 4.753215e-01