Thiamethoxam applications in squash crops

A dataset on the effect of the thiamethoxam application method and plant variety on bees.

Usage

data("thiamethoxam")

Format

A data frame with 165 observations and 11 variables:

trt

Treatment.

var

Variety.

rep

Replicate.

fruit

Average fruit number per plant.

avg_mass

Individual Fruit mass average (g).

mass

Fruit mass per plant (g).

yield

Yield (4 plants).

visit

Bee visits per plot.

foliage

Proportion of foliage consumed by striped cucumber beetle.

scb

Striped cucumber beetle per plant.

defoliation

Defoliation percentage.

Source

Obregon D, Pederson G, Taylor A, Poveda K (2022). ``The Pest Control and Pollinator Protection Dilemma: The Case of Thiamethoxam Prophylactic Applications in Squash Crops.'' PLOS ONE, 17(5), 1--18. doi:10.1371/journal.pone.0267984 .

Examples

data("thiamethoxam")
thiamethoxam
#>        trt var rep     fruit avg_mass      mass     yield visit foliage
#> 1     None  GZ   1  8.000000 241.9469  2751.130  6762.540  2.00 0.02440
#> 2     None  GZ   2  2.250000 305.4742  9672.890 13146.970  1.75 0.05240
#> 3     None  GZ   4  6.000000 493.9329  5887.710 16072.260  7.75 0.05199
#> 4     None  GZ   5  4.750000 408.8976  8804.640 16338.010  6.25 0.02001
#> 5     None  GZ   6  2.500000 346.9235  5402.390 13961.640  4.50 0.02683
#> 6     None  GZ   7 10.666667 326.6411  7230.100 10072.720  6.75 0.00246
#> 7     None SPM   1  5.750000 187.4518  2032.160  2777.700 10.50 0.01040
#> 8     None SPM   2  2.500000 267.2553  3672.120  9086.270  4.75 0.00697
#> 9     None SPM   3  5.250000 177.8564  4471.450  7931.380 11.00 0.01225
#> 10    None SPM   4  7.250000 236.8101  6152.840  9678.530  8.75 0.00106
#> 11    None SPM   5  4.750000 333.8460 11555.980 16910.440 14.75 0.00229
#> 12    None SPM   6  2.500000 401.1302  8836.410 14485.820  9.50 0.00034
#> 13    None SPM   7 11.500000 155.2796  1426.310  4981.840 13.25 0.00114
#> 14    None  GZ   1  8.000000 241.9469  2751.130  6762.540  3.75 0.00430
#> 15    None  GZ   2  2.250000 305.4742  9672.890 13146.970  8.25 0.00274
#> 16    None  GZ   3  5.666667 279.2901  7492.270 11809.390  6.25 0.00418
#> 17    None  GZ   4  6.000000 493.9329  5887.710 16072.260  7.00 0.00161
#> 18    None  GZ   5  4.750000 408.8976  8804.640 16338.010  8.25 0.00040
#> 19    None  GZ   6  2.500000 346.9235  5402.390 13961.640  6.50 0.00186
#> 20    None  GZ   7 10.666667 326.6411  7230.100 10072.720  7.75 0.00131
#> 21    None SPM   1  5.750000 187.4518  2032.160  2777.700  8.00 0.00978
#> 22    None SPM   2  2.500000 267.2553  3672.120  9086.270  9.25 0.00334
#> 23    None SPM   3  5.250000 177.8564  4471.450  7931.380  5.75 0.01450
#> 24    None SPM   4  7.250000 236.8101  6152.840  9678.530  5.75 0.01325
#> 25    None SPM   5  4.750000 333.8460 11555.980 16910.440  9.00 0.01323
#> 26    None SPM   6  2.500000 401.1302  8836.410 14485.820  8.00 0.00506
#> 27    None SPM   7 11.500000 155.2796  1426.310  4981.840  8.25 0.00580
#> 28    None  GZ   1  8.000000 241.9469  2751.130  6762.540  3.75 0.00430
#> 29    None  GZ   2  2.250000 305.4742  9672.890 13146.970  8.25 0.00274
#> 30    None  GZ   3  5.666667 279.2901  7492.270 11809.390  6.25 0.00418
#> 31    None  GZ   4  6.000000 493.9329  5887.710 16072.260  7.00 0.00161
#> 32    None  GZ   5  4.750000 408.8976  8804.640 16338.010  8.25 0.00040
#> 33    None  GZ   6  2.500000 346.9235  5402.390 13961.640  6.50 0.00186
#> 34    None  GZ   7 10.666667 326.6411  7230.100 10072.720  7.75 0.00131
#> 35    None SPM   1  5.750000 187.4518  2032.160  2777.700  8.00 0.00978
#> 36    None SPM   2  2.500000 267.2553  3672.120  9086.270  9.25 0.00334
#> 37    None SPM   3  5.250000 177.8564  4471.450  7931.380  5.75 0.01450
#> 38    None SPM   4  7.250000 236.8101  6152.840  9678.530  5.75 0.01325
#> 39    None SPM   5  4.750000 333.8460 11555.980 16910.440  9.00 0.01323
#> 40    None SPM   6  2.500000 401.1302  8836.410 14485.820  8.00 0.00506
#> 41    None SPM   7 11.500000 155.2796  1426.310  4981.840  8.25 0.00580
#> 42   Spray  GZ   1  3.250000 265.3393  2284.660  5537.660  1.00 0.02006
#> 43   Spray  GZ   2  4.250000 310.4869  2312.150  8453.250  2.25 0.01199
#> 44   Spray  GZ   3  5.000000 240.1425  1795.820  6443.040  3.50 0.02636
#> 45   Spray  GZ   4  2.500000 181.3359  2781.670  5567.520  4.25 0.07131
#> 46   Spray  GZ   5  3.750000 235.2383  2302.770  6906.860  0.75 0.05768
#> 47   Spray  GZ   6  7.250000 391.5038  2027.860 12633.230  6.50 0.01099
#> 48   Spray  GZ   7  8.000000 442.7715  6242.720 17418.230  9.00 0.00198
#> 49   Spray SPM   1  3.500000 181.9009   987.210  4805.840  5.75 0.00104
#> 50   Spray SPM   2  4.750000 298.3934  6801.760 11997.930 11.75 0.00129
#> 51   Spray SPM   3  6.750000 276.2565  4694.510 11995.360 13.00 0.00333
#> 52   Spray SPM   4  7.000000 271.9790  1433.700 10879.160 12.75 0.00014
#> 53   Spray SPM   5  6.750000 402.3045 14564.780 20852.390 14.25 0.00048
#> 54   Spray SPM   6  4.750000 235.5726  1753.863  8567.234 11.75 0.00064
#> 55   Spray SPM   7  3.000000 212.9332  2799.690  7190.840 10.00 0.00113
#> 56   Spray  GZ   1  3.250000 265.3393  2284.660  5537.660  1.00 0.02006
#> 57   Spray  GZ   2  4.250000 310.4869  2312.150  8453.250  2.25 0.01199
#> 58   Spray  GZ   3  5.000000 240.1425  1795.820  6443.040  3.50 0.02636
#> 59   Spray  GZ   4  2.500000 181.3359  2781.670  5567.520  4.25 0.07131
#> 60   Spray  GZ   5  3.750000 235.2383  2302.770  6906.860  0.75 0.05768
#> 61   Spray  GZ   6  7.250000 391.5038  2027.860 12633.230  6.50 0.01099
#> 62   Spray  GZ   7  8.000000 442.7715  6242.720 17418.230  9.00 0.00198
#> 63   Spray SPM   1  3.500000 181.9009   987.210  4805.840  2.00 0.02440
#> 64   Spray SPM   2  4.750000 298.3934  6801.760 11997.930  1.75 0.05240
#> 65   Spray SPM   4  7.000000 271.9790  1433.700 10879.160  7.75 0.05199
#> 66   Spray SPM   5  6.750000 402.3045 14564.780 20852.390  6.25 0.02001
#> 67   Spray SPM   6  4.750000 235.5726  1753.863  8567.234  4.50 0.02683
#> 68   Spray SPM   7  3.000000 212.9332  2799.690  7190.840  6.75 0.00246
#> 69   Spray  GZ   1  3.250000 265.3393  2284.660  5537.660  1.00 0.02006
#> 70   Spray  GZ   2  4.250000 310.4869  2312.150  8453.250  2.25 0.01199
#> 71   Spray  GZ   3  5.000000 240.1425  1795.820  6443.040  3.50 0.02636
#> 72   Spray  GZ   4  2.500000 181.3359  2781.670  5567.520  4.25 0.07131
#> 73   Spray  GZ   5  3.750000 235.2383  2302.770  6906.860  0.75 0.05768
#> 74   Spray  GZ   6  7.250000 391.5038  2027.860 12633.230  6.50 0.01099
#> 75   Spray  GZ   7  8.000000 442.7715  6242.720 17418.230  9.00 0.00198
#> 76   Spray SPM   1  3.500000 181.9009   987.210  4805.840  2.00 0.02440
#> 77   Spray SPM   2  4.750000 298.3934  6801.760 11997.930  1.75 0.05240
#> 78   Spray SPM   4  7.000000 271.9790  1433.700 10879.160  7.75 0.05199
#> 79   Spray SPM   5  6.750000 402.3045 14564.780 20852.390  6.25 0.02001
#> 80   Spray SPM   6  4.750000 235.5726  1753.863  8567.234  4.50 0.02683
#> 81   Spray SPM   7  3.000000 212.9332  2799.690  7190.840  6.75 0.00246
#> 82  Furrow  GZ   1  5.750000 424.1494 13956.729 19848.309  3.75 0.00430
#> 83  Furrow  GZ   2 10.250000 561.5778 15274.470 28797.650  8.25 0.00274
#> 84  Furrow  GZ   3 12.250000 512.8685 11708.250 25084.360  6.25 0.00418
#> 85  Furrow  GZ   4 10.750000 718.2468 14716.290 34789.520  7.00 0.00161
#> 86  Furrow  GZ   5 12.500000 724.2871 15498.520 34326.710  8.25 0.00040
#> 87  Furrow  GZ   6  7.750000 568.5102 11519.240 22808.510  6.50 0.00186
#> 88  Furrow  GZ   7  4.500000 229.1179  3900.290  5110.770  7.75 0.00131
#> 89  Furrow SPM   1  6.000000 323.4027  4807.740 12296.260  3.75 0.00051
#> 90  Furrow SPM   2  6.500000 361.5754  8255.500 15380.240  9.75 0.00109
#> 91  Furrow SPM   3 13.000000 333.9503 15560.070 24311.700  8.00 0.00090
#> 92  Furrow SPM   4  5.750000 376.3996  9481.420 17017.200 18.50 0.00041
#> 93  Furrow SPM   5  7.000000 344.4121 10080.080 17497.660 14.25 0.00041
#> 94  Furrow SPM   6  7.000000 344.0429  4323.400 19145.430 17.00 0.00101
#> 95  Furrow SPM   7  9.250000 376.1967 10414.190 22368.160 23.50 0.00221
#> 96  Furrow  GZ   1  5.750000 424.1494 13956.729 19848.309  5.75 0.00104
#> 97  Furrow  GZ   2 10.250000 561.5778 15274.470 28797.650 11.75 0.00129
#> 98  Furrow  GZ   3 12.250000 512.8685 11708.250 25084.360 13.00 0.00333
#> 99  Furrow  GZ   4 10.750000 718.2468 14716.290 34789.520 12.75 0.00014
#> 100 Furrow  GZ   5 12.500000 724.2871 15498.520 34326.710 14.25 0.00048
#> 101 Furrow  GZ   6  7.750000 568.5102 11519.240 22808.510 11.75 0.00064
#> 102 Furrow  GZ   7  4.500000 229.1179  3900.290  5110.770 10.00 0.00113
#> 103 Furrow SPM   1  6.000000 323.4027  4807.740 12296.260 10.50 0.01040
#> 104 Furrow SPM   2  6.500000 361.5754  8255.500 15380.240  4.75 0.00697
#> 105 Furrow SPM   3 13.000000 333.9503 15560.070 24311.700 11.00 0.01225
#> 106 Furrow SPM   4  5.750000 376.3996  9481.420 17017.200  8.75 0.00106
#> 107 Furrow SPM   5  7.000000 344.4121 10080.080 17497.660 14.75 0.00229
#> 108 Furrow SPM   6  7.000000 344.0429  4323.400 19145.430  9.50 0.00034
#> 109 Furrow SPM   7  9.250000 376.1967 10414.190 22368.160 13.25 0.00114
#> 110 Furrow  GZ   1  5.750000 424.1494 13956.729 19848.309  5.75 0.00104
#> 111 Furrow  GZ   2 10.250000 561.5778 15274.470 28797.650 11.75 0.00129
#> 112 Furrow  GZ   3 12.250000 512.8685 11708.250 25084.360 13.00 0.00333
#> 113 Furrow  GZ   4 10.750000 718.2468 14716.290 34789.520 12.75 0.00014
#> 114 Furrow  GZ   5 12.500000 724.2871 15498.520 34326.710 14.25 0.00048
#> 115 Furrow  GZ   6  7.750000 568.5102 11519.240 22808.510 11.75 0.00064
#> 116 Furrow  GZ   7  4.500000 229.1179  3900.290  5110.770 10.00 0.00113
#> 117 Furrow SPM   1  6.000000 323.4027  4807.740 12296.260 10.50 0.01040
#> 118 Furrow SPM   2  6.500000 361.5754  8255.500 15380.240  4.75 0.00697
#> 119 Furrow SPM   3 13.000000 333.9503 15560.070 24311.700 11.00 0.01225
#> 120 Furrow SPM   4  5.750000 376.3996  9481.420 17017.200  8.75 0.00106
#> 121 Furrow SPM   5  7.000000 344.4121 10080.080 17497.660 14.75 0.00229
#> 122 Furrow SPM   6  7.000000 344.0429  4323.400 19145.430  9.50 0.00034
#> 123 Furrow SPM   7  9.250000 376.1967 10414.190 22368.160 13.25 0.00114
#> 124   Seed  GZ   1  3.500000 240.4407  2322.830  8105.810  8.00 0.00978
#> 125   Seed  GZ   2  4.750000 177.1445  2911.880  5353.610  9.25 0.00334
#> 126   Seed  GZ   3  6.250000 275.5670  3815.230  7201.940  5.75 0.01450
#> 127   Seed  GZ   4  7.250000 410.1381  9363.800 15924.150  5.75 0.01325
#> 128   Seed  GZ   5  7.000000 363.3837  7600.050 13619.620  9.00 0.01323
#> 129   Seed  GZ   6  8.250000 253.8870  5728.910  8847.140  8.00 0.00506
#> 130   Seed  GZ   7  4.000000 629.9551 11189.780 27680.210  8.25 0.00580
#> 131   Seed SPM   1  1.000000 224.1095  4621.860  9094.950  1.75 0.00797
#> 132   Seed SPM   2  4.000000 102.2145  2351.730  2908.830  8.25 0.00231
#> 133   Seed SPM   3  7.000000 199.8702  3105.580  6408.560  4.50 0.00660
#> 134   Seed SPM   4  4.000000 323.7941  6456.040 13900.640  9.00 0.00040
#> 135   Seed SPM   5  7.500000 246.6138  2464.400  7454.380 10.75 0.00933
#> 136   Seed SPM   6  6.750000 219.4593  5097.980  6918.860 21.25 0.00254
#> 137   Seed SPM   7  3.250000 423.6295 16016.320 27053.110 18.50 0.00088
#> 138   Seed  GZ   1  3.500000 240.4407  2322.830  8105.810  3.75 0.00051
#> 139   Seed  GZ   2  4.750000 177.1445  2911.880  5353.610  9.75 0.00109
#> 140   Seed  GZ   3  6.250000 275.5670  3815.230  7201.940  8.00 0.00090
#> 141   Seed  GZ   4  7.250000 410.1381  9363.800 15924.150 18.50 0.00041
#> 142   Seed  GZ   5  7.000000 363.3837  7600.050 13619.620 14.25 0.00041
#> 143   Seed  GZ   6  8.250000 253.8870  5728.910  8847.140 17.00 0.00101
#> 144   Seed  GZ   7  4.000000 629.9551 11189.780 27680.210 23.50 0.00221
#> 145   Seed SPM   1  1.000000 224.1095  4621.860  9094.950  1.75 0.00797
#> 146   Seed SPM   2  4.000000 102.2145  2351.730  2908.830  8.25 0.00231
#> 147   Seed SPM   3  7.000000 199.8702  3105.580  6408.560  4.50 0.00660
#> 148   Seed SPM   4  4.000000 323.7941  6456.040 13900.640  9.00 0.00040
#> 149   Seed SPM   5  7.500000 246.6138  2464.400  7454.380 10.75 0.00933
#> 150   Seed SPM   6  6.750000 219.4593  5097.980  6918.860 21.25 0.00254
#> 151   Seed SPM   7  3.250000 423.6295 16016.320 27053.110 18.50 0.00088
#> 152   Seed  GZ   1  3.500000 240.4407  2322.830  8105.810  3.75 0.00051
#> 153   Seed  GZ   2  4.750000 177.1445  2911.880  5353.610  9.75 0.00109
#> 154   Seed  GZ   3  6.250000 275.5670  3815.230  7201.940  8.00 0.00090
#> 155   Seed  GZ   4  7.250000 410.1381  9363.800 15924.150 18.50 0.00041
#> 156   Seed  GZ   5  7.000000 363.3837  7600.050 13619.620 14.25 0.00041
#> 157   Seed  GZ   6  8.250000 253.8870  5728.910  8847.140 17.00 0.00101
#> 158   Seed  GZ   7  4.000000 629.9551 11189.780 27680.210 23.50 0.00221
#> 159   Seed SPM   1  1.000000 224.1095  4621.860  9094.950  1.75 0.00797
#> 160   Seed SPM   2  4.000000 102.2145  2351.730  2908.830  8.25 0.00231
#> 161   Seed SPM   3  7.000000 199.8702  3105.580  6408.560  4.50 0.00660
#> 162   Seed SPM   4  4.000000 323.7941  6456.040 13900.640  9.00 0.00040
#> 163   Seed SPM   5  7.500000 246.6138  2464.400  7454.380 10.75 0.00933
#> 164   Seed SPM   6  6.750000 219.4593  5097.980  6918.860 21.25 0.00254
#> 165   Seed SPM   7  3.250000 423.6295 16016.320 27053.110 18.50 0.00088
#>          scb defoliation
#> 1    1.42857       21.25
#> 2    0.37500       30.00
#> 3    0.25000       48.75
#> 4    3.87500       15.38
#> 5    2.75000       24.63
#> 6    3.12500        9.63
#> 7    0.37500        4.88
#> 8    0.12500        1.00
#> 9    1.50000        0.88
#> 10   1.25000        1.38
#> 11   0.62500        1.63
#> 12   1.75000        7.63
#> 13   3.00000        1.13
#> 14   4.75000        2.13
#> 15   1.25000        1.88
#> 16   2.00000        3.13
#> 17   1.37500        0.00
#> 18   2.00000        0.25
#> 19   2.50000        0.00
#> 20   1.25000        0.00
#> 21   5.37500       17.13
#> 22   3.87600       15.00
#> 23   2.50000        8.00
#> 24   4.00000       12.50
#> 25   1.37500       10.25
#> 26   1.25000       13.13
#> 27   5.00000       20.00
#> 28   4.75000        2.13
#> 29   1.25000        1.88
#> 30   2.00000        3.13
#> 31   1.37500        0.00
#> 32   2.00000        0.25
#> 33   2.50000        0.00
#> 34   1.25000        0.00
#> 35   5.37500       17.13
#> 36   3.87600       15.00
#> 37   2.50000        8.00
#> 38   4.00000       12.50
#> 39   1.37500       10.25
#> 40   1.25000       13.13
#> 41   5.00000       20.00
#> 42  10.12500       15.63
#> 43   5.75000        9.00
#> 44   3.62500       15.00
#> 45   3.25000       39.38
#> 46   1.87500       30.00
#> 47   2.75000        7.38
#> 48   2.75000       11.88
#> 49   5.75000        1.63
#> 50  22.87500        2.75
#> 51   2.62500        0.50
#> 52   2.00000        0.75
#> 53   1.75000        2.00
#> 54   1.50000        0.88
#> 55   5.37500        0.75
#> 56  10.12500       15.63
#> 57   5.75000        9.00
#> 58   3.62500       15.00
#> 59   3.25000       39.38
#> 60   1.87500       30.00
#> 61   2.75000        7.38
#> 62   2.75000       11.88
#> 63   1.42857       21.25
#> 64   0.37500       30.00
#> 65   0.25000       48.75
#> 66   3.87500       15.38
#> 67   2.75000       24.63
#> 68   3.12500        9.63
#> 69  10.12500       15.63
#> 70   5.75000        9.00
#> 71   3.62500       15.00
#> 72   3.25000       39.38
#> 73   1.87500       30.00
#> 74   2.75000        7.38
#> 75   2.75000       11.88
#> 76   1.42857       21.25
#> 77   0.37500       30.00
#> 78   0.25000       48.75
#> 79   3.87500       15.38
#> 80   2.75000       24.63
#> 81   3.12500        9.63
#> 82   4.75000        2.13
#> 83   1.25000        1.88
#> 84   2.00000        3.13
#> 85   1.37500        0.00
#> 86   2.00000        0.25
#> 87   2.50000        0.00
#> 88   1.25000        0.00
#> 89        NA        0.50
#> 90   2.75000        0.00
#> 91   2.25000        0.00
#> 92   0.75000        0.25
#> 93   1.87500        0.00
#> 94   0.50000        0.25
#> 95   1.12500        0.25
#> 96   5.75000        1.63
#> 97  22.87500        2.75
#> 98   2.62500        0.50
#> 99   2.00000        0.75
#> 100  1.75000        2.00
#> 101  1.50000        0.88
#> 102  5.37500        0.75
#> 103  0.37500        4.88
#> 104  0.12500        1.00
#> 105  1.50000        0.88
#> 106  1.25000        1.38
#> 107  0.62500        1.63
#> 108  1.75000        7.63
#> 109  3.00000        1.13
#> 110  5.75000        1.63
#> 111 22.87500        2.75
#> 112  2.62500        0.50
#> 113  2.00000        0.75
#> 114  1.75000        2.00
#> 115  1.50000        0.88
#> 116  5.37500        0.75
#> 117  0.37500        4.88
#> 118  0.12500        1.00
#> 119  1.50000        0.88
#> 120  1.25000        1.38
#> 121  0.62500        1.63
#> 122  1.75000        7.63
#> 123  3.00000        1.13
#> 124  5.37500       17.13
#> 125  3.87600       15.00
#> 126  2.50000        8.00
#> 127  4.00000       12.50
#> 128  1.37500       10.25
#> 129  1.25000       13.13
#> 130  5.00000       20.00
#> 131  1.00000        2.25
#> 132  2.37500        3.00
#> 133  0.62500        3.00
#> 134  1.25000        1.75
#> 135  1.12500        0.88
#> 136  1.62500        2.13
#> 137  1.50000        1.00
#> 138       NA        0.50
#> 139  2.75000        0.00
#> 140  2.25000        0.00
#> 141  0.75000        0.25
#> 142  1.87500        0.00
#> 143  0.50000        0.25
#> 144  1.12500        0.25
#> 145  1.00000        2.25
#> 146  2.37500        3.00
#> 147  0.62500        3.00
#> 148  1.25000        1.75
#> 149  1.12500        0.88
#> 150  1.62500        2.13
#> 151  1.50000        1.00
#> 152       NA        0.50
#> 153  2.75000        0.00
#> 154  2.25000        0.00
#> 155  0.75000        0.25
#> 156  1.87500        0.00
#> 157  0.50000        0.25
#> 158  1.12500        0.25
#> 159  1.00000        2.25
#> 160  2.37500        3.00
#> 161  0.62500        3.00
#> 162  1.25000        1.75
#> 163  1.12500        0.88
#> 164  1.62500        2.13
#> 165  1.50000        1.00