Empirical likelihood for general estimating functions

Computes empirical likelihood with general estimating functions.

Usage

el_eval(g, weights = NULL, control = el_control())

Arguments

g

A numeric matrix, or an object that can be coerced to a numeric matrix. Each row corresponds to an observation of an estimating function. The number of rows must be greater than the number of columns.

weights

An optional numeric vector of weights to be used in the fitting process. The length of the vector must be the same as the number of rows in g. Defaults to NULL, corresponding to identical weights. If non-NULL, weighted empirical likelihood is computed.

control

An object of class ControlEL constructed by el_control().

Value

A list with the following components:

optim

A list with the following optimization results:

lambda Lagrange multiplier of the dual problem.

iterations Number of iterations performed.

convergence Convergence status.

logp

Log probabilities obtained from empirical likelihood.

logl

Empirical log-likelihood.

loglr

Empirical log-likelihood ratio.

statistic

Minus twice the empirical log-likelihood ratio statistic that has an asymptotic chi-square distribution.

df

Degrees of freedom of the statistic.

pval

\(p\)-value of the statistic.

npar

Number of parameters.

weights

Re-scaled weights used for model fitting.

Details

el_eval evaluates empirical likelihood with a \(n\) by \(p\) numeric matrix argument g, whose \(i\)th row is \(g(X_i, \theta)\). Since the estimating function can be arbitrary, el_eval does not return an object of class EL, and the associated generics and methods are not applicable.

References

Qin J, Lawless J (1994). “Empirical Likelihood and General Estimating Equations.” The Annals of Statistics, 22(1), 300–325. doi:10.1214/aos/1176325370 .

See also

Examples

x <- rnorm(50)
par <- 0
g <- x - par
el_eval(g, weights = rep(c(1, 2), each = 25))
#> $optim
#> $optim$lambda
#> [1] 0.09667299
#> 
#> $optim$iterations
#> [1] 4
#> 
#> $optim$convergence
#> [1] TRUE
#> 
#> 
#> $logp
#>  [1] -4.291585 -4.412959 -4.333182 -4.525510 -4.474924 -4.291441 -4.358383
#>  [8] -4.519666 -4.307215 -4.269354 -4.337247 -4.177915 -4.122901 -4.224714
#> [15] -4.397993 -4.169150 -4.355237 -4.347501 -4.056720 -4.299711 -4.240229
#> [22] -4.320081 -4.430437 -4.487118 -4.258879 -3.703309 -3.725055 -3.566390
#> [29] -3.722956 -3.716538 -3.690127 -3.750382 -3.553013 -3.761916 -3.639413
#> [36] -3.524403 -3.645837 -3.277351 -3.730798 -3.658377 -3.546426 -3.646439
#> [43] -3.652876 -3.489508 -3.472939 -3.697880 -3.702478 -3.594928 -3.729234
#> [50] -3.615103
#> 
#> $logl
#> [1] -193.0916
#> 
#> $loglr
#> [1] -0.3221058
#> 
#> $statistic
#> [1] 0.6442116
#> 
#> $df
#> [1] 1
#> 
#> $p.value
#> [1] 0.4221898
#> 
#> $nobs
#> [1] 50
#> 
#> $npar
#> [1] 1
#> 
#> $weights
#>  [1] 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667
#>  [8] 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667
#> [15] 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667
#> [22] 0.6666667 0.6666667 0.6666667 0.6666667 1.3333333 1.3333333 1.3333333
#> [29] 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333
#> [36] 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333
#> [43] 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333 1.3333333
#> [50] 1.3333333
#>