Empirical likelihood for generalized linear models

Fits a generalized linear model with empirical likelihood.

Usage

el_glm(
  formula,
  family = gaussian,
  data,
  weights = NULL,
  na.action,
  control = el_control(),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  ...
)

Arguments

formula

An object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted.

family

A description of the error distribution and link function to be used in the model. Only the result of a call to a family function is supported. See ‘Details’.

data

An optional data frame, list or environment (or object coercible by as.data.frame() to a data frame) containing the variables in the formula. If not found in data, the variables are taken from environment(formula).

weights

An optional numeric vector of weights to be used in the fitting process. Defaults to NULL, corresponding to identical weights. If non-NULL, weighted empirical likelihood is computed.

na.action

A function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset.

control

An object of class ControlEL constructed by el_control().

start

Starting values for the parameters in the linear predictor. Defaults to NULL and is passed to glm.fit().

etastart

Starting values for the linear predictor. Defaults to NULL and is passed to glm.fit().

mustart

Starting values for the vector of means. Defaults to NULL and is passed to glm.fit().

...

Additional arguments to be passed to glm.control().

Value

An object of class of GLM.

Details

The available families and link functions are as follows:

gaussian: identity, log, and inverse.

bimomial: logit, probit, and log.

poisson: log, identity, and sqrt.

Included in the tests are the overall test with $$H_0: \beta_1 = \beta_2 = \cdots = \beta_{p-1} = 0,$$ and the tests for each parameter with $$H_{0j}: \beta_j = 0,\ j = 0, \dots, p-1.$$ The test results are returned as optim and parTests, respectively.

References

Chen SX, Cui H (2003). “An Extended Empirical Likelihood for Generalized Linear Models.” Statistica Sinica, 13(1), 69–81.

Kolaczyk ED (1994). “Empirical Likelihood for Generalized Linear Models.” Statistica Sinica, 4(1), 199–218.

See also

Examples

n <- 50
x <- rnorm(n)
x2 <- rnorm(n)
l <- -2 + 0.2 * x + 3 * x2
mu <- 1 / (1 + exp(-l))
y <- rbinom(n, 1, mu)
df <- data.frame(y, x, x2)
fit <- el_glm(y ~ x + x2,
  family = binomial, df, weights = NULL,
  na.action = na.omit, start = NULL, etastart = NULL, mustart = NULL
)
summary(fit)
#> 
#> Call:
#> el_glm(formula = y ~ x + x2, family = binomial, data = df, weights = NULL, 
#>     na.action = na.omit, start = NULL, etastart = NULL, mustart = NULL)
#> 
#> Coefficients:
#>             Estimate  Chisq Pr(>Chisq)    
#> (Intercept) -2.18742 15.871   6.78e-05 ***
#> x           -0.06112  0.019      0.891    
#> x2           3.62016 50.587   1.14e-12 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Chisq: 54.41, df: 2, Pr(>Chisq): 1.534e-12 
#> 
#> Constrained EL: converged 
#>