Empirical likelihood for linear models

Fits a linear model with empirical likelihood.

Usage

el_lm(formula, data, weights = NULL, na.action, control = el_control(), ...)

Arguments

formula

An object of class formula (or one that can be coerced to that class) for a symbolic description of the model to be fitted.

data

An optional data frame, list or environment (or object coercible by as.data.frame() to a data frame) containing the variables in formula. If not found in data, the variables are taken from environment(formula).

weights

An optional numeric vector of weights to be used in the fitting process. Defaults to NULL, corresponding to identical weights. If non-NULL, weighted empirical likelihood is computed.

na.action

A function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset.

control

An object of class ControlEL constructed by el_control().

...

Additional arguments to be passed to the low level regression fitting functions. See ‘Details’.

Value

An object of class of LM.

Details

Suppose that we observe \(n\) independent random variables \((X_i, Y_i)\) from a common distribution, where \(X_i\) is the \(p\)-dimensional covariate (including the intercept if any) and \(Y_i\) is the response. We consider the following linear regression model: $$Y_i = X_i^\top \theta + \epsilon_i,$$ where \(\theta = (\theta_0, \dots, \theta_{p-1})\) is an unknown \(p\)-dimensional parameter and the errors \(\epsilon_i\) are independent random variables that satisfy \(\textrm{E}(\epsilon_i | X_i)\) = 0. We assume that the errors have finite conditional variance. Then the least square estimator of \(\theta\) solves the following estimating equation: $$\sum_{i = 1}^n(Y_i - X_i^\top \theta)X_i = 0.$$ el_lm() first computes the parameter estimates by calling lm.fit() (with ... if any) with the model.frame and model.matrix obtained from the formula. Note that the maximum empirical likelihood estimator is the same as the least square estimator in our model. Next, it performs hypothesis tests based on an asymptotic chi-squared distribution of empirical likelihood ratio statistics. Included in the tests are overall test with $$H_0: \theta_1 = \theta_2 = \cdots = \theta_{p-1} = 0,$$ and significance tests for each parameter with $$H_{0j}: \theta_j = 0,\ j = 0, \dots, p-1.$$ The test results are returned as optim and sigTests, respectively.

References

Owen A (1991). “Empirical Likelihood for Linear Models.” The Annals of Statistics, 19(4), 1725–1747. doi:10.1214/aos/1176348368 .

See also

Examples

set.seed(5649)
df <- data.frame(y = rnorm(50), x = rnorm(50))
fit <- el_lm(y ~ x, df)
summary(fit)
#> 
#> Call:
#> el_lm(formula = y ~ x, data = df)
#> 
#> Coefficients:
#>             Estimate Chisq Pr(>Chisq)
#> (Intercept)  -0.1132 1.081      0.299
#> x            -0.1090 1.768      0.184
#> 
#> Chisq: 1.768, df: 1, Pr(>Chisq): 0.1837
#> 
#> Constrained EL: converged 
#> 

fit2 <- el_lm(y ~ x, df, weights = rep(c(1, 2), each = 25))
summary(fit2)
#> 
#> Call:
#> el_lm(formula = y ~ x, data = df, weights = rep(c(1, 2), each = 25))
#> 
#> Coefficients:
#>             Estimate Chisq Pr(>Chisq)
#> (Intercept)  -0.1457 1.900      0.168
#> x            -0.1056 1.568      0.211
#> 
#> Chisq: 1.568, df: 1, Pr(>Chisq): 0.2105
#> 
#> Constrained EL: converged 
#> 

df[1, 2] <- NA
fit3 <- el_lm(y ~ x, df, na.action = na.omit)
summary(fit3)
#> 
#> Call:
#> el_lm(formula = y ~ x, data = df, na.action = na.omit)
#> 
#> Coefficients:
#>             Estimate Chisq Pr(>Chisq)
#> (Intercept)  -0.1086 0.965      0.326
#> x            -0.1099 1.792      0.181
#> 
#>   (1 observation deleted due to missingness)
#> 
#> Chisq: 1.792, df: 1, Pr(>Chisq): 0.1806
#> 
#> Constrained EL: converged 
#>