Empirical likelihood for the mean

Computes empirical likelihood for the mean.

Usage

el_mean(x, par, weights = NULL, control = el_control())

Arguments

x

A numeric matrix, or an object that can be coerced to a numeric matrix. Each row corresponds to an observation. The number of rows must be greater than the number of columns.

par

A numeric vector of parameter values to be tested. The length of the vector must be the same as the number of columns in x.

weights

An optional numeric vector of weights to be used in the fitting process. The length of the vector must be the same as the number of rows in x. Defaults to NULL, corresponding to identical weights. If non-NULL, weighted empirical likelihood is computed.

control

An object of class ControlEL constructed by el_control().

Value

An object of class EL.

References

Owen A (1990). “Empirical Likelihood Ratio Confidence Regions.” The Annals of Statistics, 18(1), 90–120. doi:10.1214/aos/1176347494 .

See also

Examples

## Scalar mean
set.seed(414)
x <- rnorm(100)
par <- 0
el_mean(x, par)
#> 
#> 	Empirical Likelihood: mean 
#> 
#> Maximum EL estimates:
#> [1] -0.0006666
#> 
#> Chisq: 3.654e-05, df: 1, Pr(>Chisq): 0.9952
#> 
#> EL evaluation: converged 
#> 

## Vector mean
x <- matrix(rnorm(100), ncol = 2)
par <- c(0, 0)
el_mean(x, par)
#> 
#> 	Empirical Likelihood: mean 
#> 
#> Maximum EL estimates:
#> [1] 0.08394 0.04032
#> 
#> Chisq: 0.4832, df: 2, Pr(>Chisq): 0.7854
#> 
#> EL evaluation: converged 
#> 

## Weighted data
x <- matrix(rnorm(100), ncol = 2)
par <- c(0, 0)
w <- rep(c(1, 2), each = 25)
el_mean(x, par, w)
#> 
#> 	Weighted Empirical Likelihood: mean 
#> 
#> Maximum EL estimates:
#> [1] -0.06031  0.17921
#> 
#> Chisq: 1.48, df: 2, Pr(>Chisq): 0.477
#> 
#> EL evaluation: converged 
#>