Empirical likelihood multiple tests

Tests multiple linear hypotheses simultaneously.

# S4 method for EL
elmt(object, rhs = NULL, lhs = NULL, alpha = 0.05, control = el_control())

Arguments

object

An object that inherits from EL.

rhs

A numeric vector (column matrix) or a list of numeric vectors for the right-hand sides of hypotheses. Defaults to `NULL`. See ‘Details’.

lhs

A list or a numeric matrix for the left-hand sides of hypotheses. For a list `lhs`, each element must be specified as a single instance of the `lhs` in [elt()]. For a matrix `lhs`, each row gives a linear combination of the parameters in `object`. The number of columns must be equal to the number of parameters. Defaults to `NULL`. See ‘Details’.

alpha

A single numeric for the overall significance level. Defaults to `0.05`.

control

An object of class ControlEL constructed by [el_control()].

Value

An object of class of ELMT.

Details

[elmt()] tests multiple hypotheses simultaneously. Each hypothesis corresponds to the constrained empirical likelihood ratio described in CEL. `rhs` and `lhs` cannot be both `NULL`. The right-hand side and left-hand side of each hypothesis must be specified as described in [elt()].

For specifying linear contrasts more conveniently, `rhs` and `lhs` also take a numeric vector and a numeric matrix, respectively. Each element of `rhs` and each row of `lhs` correspond to a contrast (hypothesis).

The vector of empirical likelihood ratio statistics asymptotically follows a multivariate chi-square distribution under the complete null hypothesis. The multiple testing procedure asymptotically controls the family-wise error rate at the level `alpha`. Based on the distribution of the maximum of the test statistics, the adjusted p-values are estimated by Monte Carlo simulation.

References

Kim E, MacEachern S, Peruggia M (2021). “Empirical Likelihood for the Analysis of Experimental Designs.” arxiv:2112.09206. URL <https://arxiv.org/abs/2112.09206>.

See also

Examples

## Bivariate mean (list `rhs` & no `lhs`)
set.seed(143)
data("women")
fit <- el_mean(women, par = c(65, 135))
rhs <- list(c(64, 133), c(66, 140))
elmt(fit, rhs = rhs)
#> 
#> 	Empirical Likelihood Multiple Tests
#> 
#> Overall significance level: 0.05 
#> 
#> Calibration: Multivariate chi-square 
#> 
#> Hypotheses:
#>   Chisq Df p.adj
#> 1 2.069  2 0.355
#> 2 1.255  2 0.534
#> 
#> Common critical value: 5.984 
#> 

## Pairwise comparison (no `rhs` & list `lhs`)
data("clothianidin")
fit2 <- el_lm(clo ~ -1 + trt, clothianidin)
lhs2 <- list(
  "trtNaked - trtFungicide",
  "trtFungicide - trtLow",
  "trtLow - trtHigh"
)
elmt(fit2, lhs = lhs2)
#> 
#> 	Empirical Likelihood Multiple Tests
#> 
#> Overall significance level: 0.05 
#> 
#> Calibration: Multivariate chi-square 
#> 
#> Hypotheses:
#>                             Estimate Chisq Df p.adj  
#> trtNaked - trtFungicide = 0  -1.0525 5.510  1 0.052 .
#> trtFungicide - trtLow = 0    -0.6269 1.062  1 0.613  
#> trtLow - trtHigh = 0         -1.4932 3.774  1 0.135  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Common critical value: 5.583 
#> 

## Arbitrary hypotheses (list `rhs` & list `lhs`)
data("mtcars")
fit3 <- el_lm(mpg ~ wt + qsec, data = mtcars)
lhs3 <- list(c(1, 4, 0), rbind(c(0, 1, 0), c(0, 0, 1)))
rhs3 <- list(0, c(-6, 1))
elmt(fit3, rhs = rhs3, lhs = lhs3)
#> 
#> 	Empirical Likelihood Multiple Tests
#> 
#> Overall significance level: 0.05 
#> 
#> Calibration: Multivariate chi-square 
#> 
#> Hypotheses:
#>   Chisq Df p.adj
#> 1 0.037  1 0.985
#> 2 2.790  2 0.251
#> 
#> Common critical value: 6.025 
#>