Thiamethoxam applications in squash crops

A dataset on the effect of the thiamethoxam application method and plant variety on bees.

data("thiamethoxam")

Format

A data frame with 55 observations and 11 variables:

trt

Treatment.

var

Variety.

rep

Replicate.

fruit

Average fruit number per plant.

avg_mass

Individual Fruit mass average (g).

mass

Fruit mass per plant (g).

yield

Yield (4 plants).

visit

Bee visits per plot.

foliage

Proportion of foliage consumed by striped cucumber beetle.

scb

Striped cucumber beetle per plant.

defoliation

Defoliation percentage.

Source

Obregon D, Pederson G, Taylor A, Poveda K (2022). “The Pest Control and Pollinator Protection Dilemma: The Case of Thiamethoxam Prophylactic Applications in Squash Crops.” PLOS ONE, 17(5), 1--18. doi:10.1371/journal.pone.0267984 .

Examples

data("thiamethoxam")
thiamethoxam
#>       trt var rep    fruit avg_mass      mass     yield visit foliage      scb
#> 1   Spray  GZ   1  3.25000 265.3393  2284.660  5537.660  1.00 0.02006 10.12500
#> 2   Spray  GZ   2  4.25000 310.4869  2312.150  8453.250  2.25 0.01199  5.75000
#> 3   Spray  GZ   3  5.00000 240.1425  1795.820  6443.040  3.50 0.02636  3.62500
#> 4   Spray  GZ   4  2.50000 181.3359  2781.670  5567.520  4.25 0.07131  3.25000
#> 5   Spray  GZ   5  3.75000 235.2383  2302.770  6906.860  0.75 0.05768  1.87500
#> 6   Spray  GZ   6  7.25000 391.5038  2027.860 12633.230  6.50 0.01099  2.75000
#> 7   Spray  GZ   7  8.00000 442.7715  6242.720 17418.230  9.00 0.00198  2.75000
#> 8   Spray SPM   1  3.50000 181.9009   987.210  4805.840  5.75 0.00104  5.75000
#> 9   Spray SPM   2  4.75000 298.3934  6801.760 11997.930 11.75 0.00129 22.87500
#> 10  Spray SPM   3  6.75000 276.2565  4694.510 11995.360 13.00 0.00333  2.62500
#> 11  Spray SPM   4  7.00000 271.9790  1433.700 10879.160 12.75 0.00014  2.00000
#> 12  Spray SPM   5  6.75000 402.3045 14564.780 20852.390 14.25 0.00048  1.75000
#> 13  Spray SPM   6  4.75000 235.5726  1753.863  8567.234 11.75 0.00064  1.50000
#> 14  Spray SPM   7  3.00000 212.9332  2799.690  7190.840 10.00 0.00113  5.37500
#> 15 Furrow  GZ   1  5.75000 424.1494 13956.729 19848.309  3.75 0.00430  4.75000
#> 16 Furrow  GZ   2 10.25000 561.5778 15274.470 28797.650  8.25 0.00274  1.25000
#> 17 Furrow  GZ   3 12.25000 512.8685 11708.250 25084.360  6.25 0.00418  2.00000
#> 18 Furrow  GZ   4 10.75000 718.2468 14716.290 34789.520  7.00 0.00161  1.37500
#> 19 Furrow  GZ   5 12.50000 724.2871 15498.520 34326.710  8.25 0.00040  2.00000
#> 20 Furrow  GZ   6  7.75000 568.5102 11519.240 22808.510  6.50 0.00186  2.50000
#> 21 Furrow  GZ   7  4.50000 229.1179  3900.290  5110.770  7.75 0.00131  1.25000
#> 22 Furrow SPM   1  6.00000 323.4027  4807.740 12296.260  3.75 0.00051       NA
#> 23 Furrow SPM   2  6.50000 361.5754  8255.500 15380.240  9.75 0.00109  2.75000
#> 24 Furrow SPM   3 13.00000 333.9503 15560.070 24311.700  8.00 0.00090  2.25000
#> 25 Furrow SPM   4  5.75000 376.3996  9481.420 17017.200 18.50 0.00041  0.75000
#> 26 Furrow SPM   5  7.00000 344.4121 10080.080 17497.660 14.25 0.00041  1.87500
#> 27 Furrow SPM   6  7.00000 344.0429  4323.400 19145.430 17.00 0.00101  0.50000
#> 28 Furrow SPM   7  9.25000 376.1967 10414.190 22368.160 23.50 0.00221  1.12500
#> 29   None  GZ   1  8.00000 241.9469  2751.130  6762.540  2.00 0.02440  1.42857
#> 30   None  GZ   2  2.25000 305.4742  9672.890 13146.970  1.75 0.05240  0.37500
#> 31   None  GZ   4  6.00000 493.9329  5887.710 16072.260  7.75 0.05199  0.25000
#> 32   None  GZ   5  4.75000 408.8976  8804.640 16338.010  6.25 0.02001  3.87500
#> 33   None  GZ   6  2.50000 346.9235  5402.390 13961.640  4.50 0.02683  2.75000
#> 34   None  GZ   7 10.66667 326.6411  7230.100 10072.720  6.75 0.00246  3.12500
#> 35   None SPM   1  5.75000 187.4518  2032.160  2777.700 10.50 0.01040  0.37500
#> 36   None SPM   2  2.50000 267.2553  3672.120  9086.270  4.75 0.00697  0.12500
#> 37   None SPM   3  5.25000 177.8564  4471.450  7931.380 11.00 0.01225  1.50000
#> 38   None SPM   4  7.25000 236.8101  6152.840  9678.530  8.75 0.00106  1.25000
#> 39   None SPM   5  4.75000 333.8460 11555.980 16910.440 14.75 0.00229  0.62500
#> 40   None SPM   6  2.50000 401.1302  8836.410 14485.820  9.50 0.00034  1.75000
#> 41   None SPM   7 11.50000 155.2796  1426.310  4981.840 13.25 0.00114  3.00000
#> 42   Seed  GZ   1  3.50000 240.4407  2322.830  8105.810  8.00 0.00978  5.37500
#> 43   Seed  GZ   2  4.75000 177.1445  2911.880  5353.610  9.25 0.00334  3.87600
#> 44   Seed  GZ   3  6.25000 275.5670  3815.230  7201.940  5.75 0.01450  2.50000
#> 45   Seed  GZ   4  7.25000 410.1381  9363.800 15924.150  5.75 0.01325  4.00000
#> 46   Seed  GZ   5  7.00000 363.3837  7600.050 13619.620  9.00 0.01323  1.37500
#> 47   Seed  GZ   6  8.25000 253.8870  5728.910  8847.140  8.00 0.00506  1.25000
#> 48   Seed  GZ   7  4.00000 629.9551 11189.780 27680.210  8.25 0.00580  5.00000
#> 49   Seed SPM   1  1.00000 224.1095  4621.860  9094.950  1.75 0.00797  1.00000
#> 50   Seed SPM   2  4.00000 102.2145  2351.730  2908.830  8.25 0.00231  2.37500
#> 51   Seed SPM   3  7.00000 199.8702  3105.580  6408.560  4.50 0.00660  0.62500
#> 52   Seed SPM   4  4.00000 323.7941  6456.040 13900.640  9.00 0.00040  1.25000
#> 53   Seed SPM   5  7.50000 246.6138  2464.400  7454.380 10.75 0.00933  1.12500
#> 54   Seed SPM   6  6.75000 219.4593  5097.980  6918.860 21.25 0.00254  1.62500
#> 55   Seed SPM   7  3.25000 423.6295 16016.320 27053.110 18.50 0.00088  1.50000
#>    defoliation
#> 1        15.63
#> 2         9.00
#> 3        15.00
#> 4        39.38
#> 5        30.00
#> 6         7.38
#> 7        11.88
#> 8         1.63
#> 9         2.75
#> 10        0.50
#> 11        0.75
#> 12        2.00
#> 13        0.88
#> 14        0.75
#> 15        2.13
#> 16        1.88
#> 17        3.13
#> 18        0.00
#> 19        0.25
#> 20        0.00
#> 21        0.00
#> 22        0.50
#> 23        0.00
#> 24        0.00
#> 25        0.25
#> 26        0.00
#> 27        0.25
#> 28        0.25
#> 29       21.25
#> 30       30.00
#> 31       48.75
#> 32       15.38
#> 33       24.63
#> 34        9.63
#> 35        4.88
#> 36        1.00
#> 37        0.88
#> 38        1.38
#> 39        1.63
#> 40        7.63
#> 41        1.13
#> 42       17.13
#> 43       15.00
#> 44        8.00
#> 45       12.50
#> 46       10.25
#> 47       13.13
#> 48       20.00
#> 49        2.25
#> 50        3.00
#> 51        3.00
#> 52        1.75
#> 53        0.88
#> 54        2.13
#> 55        1.00